+++ Common/File/FileDescriptor.cpp
@@ -83,8 +83,8 @@ size_t Write(int fd, const std::string &str) {
 
 bool WaitUntilReady(int fd, double timeout, bool for_write) {
 	struct timeval tv;
-	tv.tv_sec = floor(timeout);
-	tv.tv_usec = (timeout - floor(timeout)) * 1000000.0;
+	tv.tv_sec = (long)floor(timeout);
+	tv.tv_usec = (long)((timeout - floor(timeout)) * 1000000.0);
 
 	fd_set fds;
 	FD_ZERO(&fds);
